Use a integer index into the member array instead of pointers.
Previously it was assumed that the amount of members that could be
present in the datatype could be determined before finalize. This is no
longer the case as, due to repacing namespaces with composite types,
types would not be available at this point. Because of this we have to
stepwise analyze each member (and sub member and so on) as their types
becomes available. In order to facilitate dynamic growth of member list,
we need an indexing strategy that does not rely on the exact position of
the member in memory, as that address can change (realloc).

This change was also motivated to have a simple index that slots can be
tagged with to indicate what member they represent.